Add description of utvalg
@arnfinn @Tovjoh I think in the helseatlas app we should have some description text.   (see issue 150 in helseatlas/helseatlas)  I have added just one now for testing.

Based on how the data is structured now, the easiest way is to add an additional variable to the existing file and use this to display the description.  However, it may be a better design to have this info (along with other info that are on the utvalg level, not opptaksområde level, such as level1_name, level2_name, type, numerator_name, denominator_name, and period) to be in a separate file.